+++ packages/app-core/src/index.ts
@@ -5,3 +5,4 @@ export * from './JBrowseConfig'
 export * from './JBrowseModel'
 export * from './Assemblies'
 export * from './RootMenu'
+export * from './menus'
