+++ packages/web-core/src/index.ts
@@ -1,3 +1,2 @@
 export * from './SessionConnections'
 export * from './BaseWebSession'
-export * from './menus'
